#38 style,type 분리 및 Toast svg 파일 import 방식으로 변경
