+++ src/components/Button/Button.tsx
@@ -1,54 +1,14 @@
 import * as React from 'react';
 import clsx from 'clsx';
-
-export interface ButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  variant?: 'contained' | 'outlined' | 'text';
-  color?: 'primary' | 'secondary';
-  size?: 'small' | 'medium' | 'large';
-}
-
-const baseStyles = 'font-medium rounded';
-
-const sizeStyles = {
-  small: 'text-xs px-2.5 py-1.5',
-  medium: 'text-sm leading-4 px-3 py-2',
-  large: 'text-sm px-4 py-2',
-};
-
-const variantStyles = {
-  contained: 'border border-transparent',
-  outlined: 'bg-transparent border',
-  text: 'bg-transparent border-transparent',
-};
-
-const colorStyles = {
-  text: {
-    primary:
-      'text-black bg-transparent hover:bg-primary-50 active:bg-primary-100',
-    secondary:
-      'text-black bg-transparent hover:bg-secondary-50 active:bg-secondary-100',
-  },
-  contained: {
-    primary:
-      'text-white bg-primary-default hover:bg-primary-600 active:bg-primary-500 rounded-lg',
-    secondary:
-      'text-white bg-secondary-default hover:bg-secondary-600 active:bg-secondary-500 rounded-lg',
-  },
-  outlined: {
-    primary:
-      'text-primary-default border-primary-default hover:bg-primary-50 active:bg-primary-100',
-    secondary:
-      'text-secondary-500 border-secondary-500 hover:bg-secondary-50 active:bg-secondary-100',
-  },
-};
-
-const disabledSyles = 'cursor-default pointer-events-none';
-const disabledVariantSyles = {
-  contained: 'text-gray-300 bg-gray-100',
-  outlined: 'text-gray-300 border-gray-300',
-  text: 'text-gray-300 bg-gray-100',
-};
+import {
+  baseStyles,
+  sizeStyles,
+  variantStyles,
+  colorStyles,
+  disabledStyles,
+  disabledVariantStyles,
+} from './Button.style';
+import { ButtonProps } from './Button.type';
 
 export function Button({
   className,
@@ -65,8 +25,8 @@ export function Button({
         baseStyles,
         variantStyles[variant],
         sizeStyles[size],
-        disabledSyles,
-        disabledVariantSyles[variant]
+        disabledStyles,
+        disabledVariantStyles[variant]
       )
     : clsx(
         className,
