+++ src/components/Button/Button.type.ts
@@ -1,0 +1,6 @@
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  variant?: 'contained' | 'outlined' | 'text';
+  color?: 'primary' | 'secondary';
+  size?: 'small' | 'medium' | 'large';
+}
