+++ src/components/Chip/Chip.style.ts
@@ -1,0 +1,14 @@
+export const baseStyles = 'font-medium rounded  h-9 rounded-full px-3 pt-1 w-fit';
+
+export const disabledStyles = 'cursor-default pointer-events-none';
+
+export const colorStyles = {
+  contained: {
+    primary: 'text-white bg-primary-500',
+    secondary: 'text-white bg-secondary-500',
+  },
+  outlined: {
+    primary: 'text-primary-500 bg-white border border-primary-500',
+    secondary: 'text-secondary-500 bg-white border border-secondary-500',
+  },
+};
