+++ src/components/Chip/Chip.type.ts
@@ -1,0 +1,7 @@
+export interface ChipProps extends Omit<React.HTMLAttributes<HTMLDivElement>, 'size'> {
+    variant?: 'contained' | 'outlined';
+    removable?: boolean;
+    color?: 'primary' | 'secondary';
+    disabled?: boolean;
+    text?: string;
+  }
