+++ src/components/Toast/Toast.type.ts
@@ -1,0 +1,7 @@
+export interface ToastProps
+  extends Omit<React.HTMLAttributes<HTMLDivElement>, 'size'> {
+  variant?: 'contained' | 'outlined';
+  type?: 'success' | 'error' | 'info' | 'warning';
+  size?: 'small' | 'medium' | 'large';
+  open?: boolean;
+}
